+++ installers/plcc/install.bash
@@ -3,6 +3,6 @@
 set -euo pipefail
 
 PLCC_HOME="${HOME}/.local/lib/plcc"
-git clone https://github.com/ourPLCC/plcc.git "${PLCC_HOME}"
+git clone -b "${PLCC_GIT_BRANCH:-main}" https://github.com/ourPLCC/plcc.git "${PLCC_HOME}"
 echo "# Add the following line to the end of your .bashrc or .zshrc"
 "${PLCC_HOME}/installers/plcc/rchook.bash"
