--- src/parse
@@ -1,49 +1,0 @@
-#!/bin/bash
-
-JACKSON_VERSION="2.15.2"
-LIBPLCC_LIB="$(cd -- "${LIBPLCC}/.." &> /dev/null ; pwd)"
-JACKSON_ANNOTATIONS="${LIBPLCC_LIB}/lib/jackson/jackson-annotations-${JACKSON_VERSION}.jar"
-JACKSON_CORE="${LIBPLCC_LIB}/lib/jackson/jackson-core-${JACKSON_VERSION}.jar"
-JACKSON_DATABIND="${LIBPLCC_LIB}/lib/jackson/jackson-databind-${JACKSON_VERSION}.jar"
-CP="${JACKSON_ANNOTATIONS}:${JACKSON_CORE}:${JACKSON_DATABIND}"
-
-if [[ -z "${CLASSPATH:-}" ]] ; then
-    CLASSPATH="./Java:${CP}"
-    export CLASSPATH
-else
-    CLASSPATH="./Java:${CLASSPATH}:${CP}"
-fi
-
-[ -d Java ] || {
-    echo "Java directory missing" >&2
-    exit 1
-}
-
-[ -f Java/Parse.class ] || {
-    echo "Java/Parse.class missing" >&2
-    exit 2
-}
-
-if  echo "$*" | grep -- "--json_ast" &> /dev/null
-then
-
-    ARGS=()
-    while [ $# -gt 0 ] ; do
-        case "$1" in
-            --json_ast)
-                shift
-            ;;
-            *)
-                ARGS+=("$1")
-                shift
-            ;;
-        esac
-    done
-
-    if ! java ParseJsonAst "${ARGS[@]}" ; then
-        >&2 echo "Did you forget to pass --json_ast to plccmk?"
-        exit 1
-    fi
-else
-    java Parse $*
-fi
