--- src/plcc
@@ -1,12 +1,0 @@
-#!/bin/bash
-
-LIB="${LIBPLCC:-/home/fossumtv/PL/src}"
-PYTHON3=python3
-PLCC="$LIB/plcc.py"
-
-[ -d "$LIB" ] || {
-        echo "$LIB: no such directory" >&2
-        exit 1
-}
-
-$PYTHON3 "$PLCC" "$@"
