+++ src/plcc/bin/plcc
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+source "$(dirname -- "${BASH_SOURCE[0]}" )/common.bash" && init
+
+python3 -m plcc "$@"
