+++ src/plcc/lib/Std/ILazy.java
