+++ src/plcc/lib/Std/Scan.java
