+++ src/plcc/lib/Std/Token.pattern
