--- src/plccmk
@@ -1,51 +1,0 @@
-#!/bin/bash
-
-LIB="${LIBPLCC:-/usr/local/pub/plcc/PLCC}"
-PYTHON3=python3
-PLCC="$LIB/plcc.py"
-
-JACKSON_VERSION="2.15.2"
-LIBPLCC_LIB="$(cd -- "${LIBPLCC}/.." &> /dev/null ; pwd)"
-JACKSON_ANNOTATIONS="${LIBPLCC_LIB}/lib/jackson/jackson-annotations-${JACKSON_VERSION}.jar"
-JACKSON_CORE="${LIBPLCC_LIB}/lib/jackson/jackson-core-${JACKSON_VERSION}.jar"
-JACKSON_DATABIND="${LIBPLCC_LIB}/lib/jackson/jackson-databind-${JACKSON_VERSION}.jar"
-CP="${JACKSON_ANNOTATIONS}:${JACKSON_CORE}:${JACKSON_DATABIND}"
-
-if [[ -z "${CLASSPATH:-}" ]] ; then
-        CLASSPATH="${CP}"
-        export CLASSPATH
-else
-        CLASSPATH="${CLASSPATH}:${CP}"
-fi
-
-if [ "$1" = "-c" ];
-then
-        rm Java/*.java Java/*.class &>/dev/null
-        shift
-fi
-
-if [ $# -gt 0 ];
-then
-        FILES="$@"
-else
-        DEFAULT=spec
-        [ -f "$DEFAULT" ] || DEFAULT=grammar
-        FILES="${FILE:-$DEFAULT}"
-fi
-
-[ -d "$LIB" ] || {
-        echo "$LIB: no such directory" >&2
-        exit 1
-}
-
-$PYTHON3 "$PLCC" $FILES || {
-        echo "cannot compile $FILES" >&2
-        exit 4
-}
-
-[ -d Java ] || {
-        echo "Java directory missing" >&2
-        exit 5
-}
-
-(cd ./Java ; javac *.java)
