--- src/rep
@@ -1,20 +1,0 @@
-#!/bin/bash
-
-if [[ -z "${CLASSPATH:-}" ]] ; then
-    CLASSPATH="./Java"
-    export CLASSPATH
-else
-    CLASSPATH="./Java:${CLASSPATH}"
-fi
-
-[ -d "${PWD}/Java" ] || {
-    echo "Java directory missing" >&2
-    exit 1
-}
-
-[ -f "${PWD}/Java/Rep.class" ] || {
-    echo "Java/Rep.class missing" >&2
-    exit 2
-}
-
-java Rep $*
