--- src/scan
@@ -1,21 +1,0 @@
-#!/bin/bash
-
-if [[ -z "${CLASSPATH:-}" ]] ; then
-    CLASSPATH="./Java"
-    export CLASSPATH
-else
-    CLASSPATH="./Java:${CLASSPATH}"
-fi
-
-
-[ -d Java ] || {
-    echo "Java directory missing" >&2
-    exit 1
-}
-
-[ -f Java/Scan.class ] || {
-    echo "Java/Scan.class missing" >&2
-    exit 2
-}
-
-java Scan
