+++ tests/parse/parses/given-grammar.plcc
@@ -1,0 +1,6 @@
+A 'A'
+B 'B'
+skip OTHER '.'
+%
+<p> ::= <aaa> B
+<aaa> **= A
