+++ tests/parse/parses/given-program.lang
@@ -1,0 +1,1 @@
+A asdf A fdsa A B
