+++ tests/parse/parses/parses.bats
@@ -1,0 +1,10 @@
+#!/usr/bin/env bats
+
+load '../../relocate_to_temp.bash'
+
+@test "PLCC parses." {
+  relocate_to_temp
+
+  plccmk -c given-grammar.plcc
+  # parse -n < given-program.lang
+}
